# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SQL Practice API
  version: "1.0.0"
  description: |
    API for the SQL Practice app (interactive SQL tasks checked against a real
    PostgreSQL database, sessions/learning plans, login/role system).

    **This spec is hand-maintained, not generated from code.** It documents a
    first, deliberately partial slice of the API (see "Phase 1" scope below) —
    if anything here looks stale compared to the actual route code, trust the
    code and update this file.

    ### Authentication
    This API uses a session cookie (`connect.sid`, set by `express-session` +
    `connect-pg-simple`), **not** bearer tokens. To use "Try it out" below:
    1. Call `POST /api/auth/login` with a real username/password — a valid
       session cookie is stored by your browser automatically.
    2. Every subsequent "Try it out" call from this same page reuses that
       cookie automatically — no token to copy/paste anywhere.
    3. Call `POST /api/auth/logout` to end the session.

    This only works because Swagger UI is served by the backend itself
    (same origin as `/api/*`) — see `src/routes/docs.js`.

    ### Phase 1 scope
    Only a first, "safe to explore" slice of the API is documented here:
    health, auth, datasets, tasks (read + solution), the query sandbox, and
    the table browser. Endpoints with heavier role/ownership matrices
    (sessions, users, progress, mentor assignments, mentor students) are
...
    full API surface in the meantime.
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: >-
        Session cookie set by POST /api/auth/login. httpOnly, sameSite=lax,
        secure in production. There is no bearer-token alternative.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required: [error]
      example:
        error: "Not authenticated."
...
    Task:
      type: object
      description: >-
        A practice task with its `solution` field stripped out. The solution
        is only ever exposed via GET /api/tasks/{id}/solution.
      properties:
        id:
          type: integer
        datasetKey:
          type: string
        category:
          type: string
        topicId:
          type: string
          nullable: true
        projectId:
          type: string
          nullable: true
          description: Present instead of topicId for Practice Project tasks.
        levelId:
          type: string
          enum: [introduction, beginner, intermediate, advanced, expert]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        title:
          type: string
        description:
          type: string
        hint:
          type: string
        tables:
          type: array
          items:
            type: string
        requiresOrderBy:
          type: boolean
          nullable: true
        validationMode:
          type: string
          enum: [strict, result_only]
          nullable: true
      example:
        id: 1
        datasetKey: academic
        category: SELECT basics
        topicId: select
        levelId: introduction
        difficulty: easy
        title: Select all students
        description: Show all columns and all rows from the students table.
        hint: Use SELECT * and FROM students.
        tables: [students]
...
    SampleDifference:
      type: object
      description: One differing row, capped at 3 per response (collectDiffs in resultComparator.js).
      properties:
        row:
          type: integer
          description: 1-based row index in the compared (sorted, if order doesn't matter) result set.
        columns:
          type: array
          items: { type: string }
          description: Lowercased column names that differ on this row.
        userValues:
          type: array
          items: {}
          description: The caller's values for `columns`, in the same order.
        expectedValues:
          type: array
          items: {}
          description: The solution's values for `columns`, in the same order.
    CheckAnswerResult:
      type: object
      description: >-
        The 200 response of POST /api/tasks/{id}/check. A single shared shape
        rather than oneOf per failureReason — the fields that vary are all
        optional, and `isCorrect`/`solutionHasJoin`/`userResult`/
        `expectedRowCount` are always present regardless of outcome. See the
        endpoint's examples for the concrete shape of each failureReason.
      required: [isCorrect, solutionHasJoin, userResult, expectedRowCount]
      properties:
        isCorrect:
          type: boolean
        failureReason:
          type: string
          enum:
            - column_count_mismatch
            - column_name_mismatch
            - row_count_mismatch
            - order_mismatch
            - duplicate_rows_mismatch
            - value_mismatch
            - query_logic_mismatch
          description: >-
            Key absent entirely (not just null) when isCorrect is true.
            query_logic_mismatch is a distinct second pass (structural/ORDER
            BY checks) that only runs after the result sets already matched —
            see the endpoint description.
        userColumns:
          type: array
          items: { type: string }
          description: Only present for column_count_mismatch / column_name_mismatch.
        expectedColumns:
          type: array
          items: { type: string }
          description: Only present for column_count_mismatch / column_name_mismatch.
        missingColumns:
          type: array
          items: { type: string }
          description: Only present for column_count_mismatch — expected columns the caller's query is missing.
        extraColumns:
          type: array
          items: { type: string }
          description: Only present for column_count_mismatch — columns the caller's query has that the solution doesn't.
        userRowCount:
          type: integer
          description: Only present for row_count_mismatch / order_mismatch / duplicate_rows_mismatch / value_mismatch.
        sampleDifferences:
          type: array
          items: { $ref: "#/components/schemas/SampleDifference" }
          description: Only present for order_mismatch / value_mismatch — up to 3 sample rows.
        logicMismatchReason:
          type: string
          enum:
            - missing_order_by
            - extra_where_condition
            - missing_where_condition
            - condition_mismatch
            - extra_distinct
            - extra_order_by
            - extra_limit
          description: Only present for failureReason=query_logic_mismatch.
        logicMismatchHint:
          type: string
          description: A human-readable hint paired with logicMismatchReason. Only present alongside it.
        solutionHasJoin:
          type: boolean
          description: Always present — whether the solution SQL contains JOIN (case-insensitive substring check).
        userResult:
          type: object
          description: Always present — the caller's own query result, echoed back for the UI to render.
          properties:
            rows: { type: array, items: { type: object } }
            columns: { type: array, items: { type: string } }
            rowCount: { type: integer }
          required: [rows, columns, rowCount]
        expectedRowCount:
          type: integer
          description: Always present — the solution's row count.
security:
  - sessionCookie: []
paths:
...
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List tasks (optionally filtered)
      description: >-
        Returns tasks with their `solution` field stripped out. All filters
        are optional and combine with AND.
      operationId: listTasks
      security: []
      x-visibility: public
      parameters:
        - name: datasetKey
          in: query
          schema: { type: string }
          example: academic
        - name: category
          in: query
          schema: { type: string }
        - name: topicId
          in: query
          schema: { type: string }
          example: select
        - name: levelId
          in: query
          schema:
            type: string
            enum: [introduction, beginner, intermediate, advanced, expert]
        - name: projectId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Matching tasks.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get a single task by id
      description: Returns the task with its `solution` field stripped out.
      operationId: getTask
      security: []
      x-visibility: public
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          example: 1
      responses:
        "200":
          description: The task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "404":
          description: No task with this id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: "Task not found"
  /api/tasks/{id}/solution:
    get:
      tags: [Tasks]
      summary: Get a task's solution SQL
      description: >-
        Requires login, but any authenticated role (admin/mentor/student) may
        fetch a solution — there is no additional role gate. This is the
        **only** route that ever returns solution SQL; everywhere else only
        derived booleans (e.g. `solutionHasJoin`) are exposed.
      operationId: getTaskSolution
      x-visibility: authenticated
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          example: 1
      responses:
        "200":
          description: The task's solution SQL.
          content:
            application/json:
              schema:
                type: object
                properties:
                  solution: { type: string }
                required: [solution]
              example:
                solution: "SELECT *\nFROM students;"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: No task with this id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/tasks/{id}/check:
    post:
      tags: [Tasks]
      summary: Check Answer — run SQL and compare it against the task's solution
      description: >-
        Runs `userSql` and the task's trusted solution SQL in parallel, then
        compares the two result sets. This is the "Check Answer" button —
        distinct from **Run Query** (POST /api/query), which only executes
        SQL and never compares it to anything or updates progress.


        **Request/response order is not what you'd expect — read this before
        testing 401 in Swagger.** The checks run in this exact order: (1)
        task lookup — 404 if missing; (2) `userSql` presence — 400 if
        missing/blank; (3) authentication — 401. **A request with no
        `userSql` gets 400, not 401, even if you are logged out** — you must
        supply a non-empty `userSql` to actually observe the 401 case.


        **sessionId behavior is more permissive than other routes, and does
        NOT return 403 for an unauthorized session — read carefully:**
        - Omitted (or empty/non-numeric): falls back to the caller's own
          first non-archived session, same as most other session-aware
          routes.
        - A `sessionId` the caller does not own: **silently resolves to "no
          session" (not an error, not a fallback to the caller's own
          session)** — the check still runs and a result is still returned,
          but nothing is persisted (see below) and the completed/archived
          guard below never even runs, because it only runs when a session
          was actually resolved. There is deliberately no 403 for "session
          not yours" on this route, unlike GET /api/sessions or
          PATCH /api/sessions/{id}.
        - A `sessionId` the caller DOES own, that is completed or archived:
          403, with the same messages PATCH-guarded routes use elsewhere.


        **Attempt recording / progress side effects — only when a session
        was actually resolved** (see above): every code path (safety
        rejection, execution error, row-limit rejection, correct, incorrect)
        writes one row to `task_attempts` (via `saveCheckAttempt`) AND
        upserts `user_task_progress` for `(user, session, task)` — status
        becomes `solved` on a correct answer (setting `solved_at` the first
        time only) or `in_progress` on anything else (including safety/
        execution failures), `attempts_count` increments, `last_attempt_at`/
        `last_submitted_sql` update. **If no session was resolved (omitted
        with no existing session, or an unauthorized sessionId as above),
        neither table is written at all** — the response is still generated,
        it just isn't persisted anywhere.


        **Dataset scope**: the task must belong to the same dataset as the
        resolved session (400 if not) — when no session is resolved, this
        falls back to checking against the academic dataset.


        **Correctness pipeline, in order:** (1) `validateSqlSafety` — SELECT/
        WITH only, keyword blocklist (403 on failure); (2)
        `validateSchemaScope` — blocks cross-dataset schema references (403);
        (3) execute both queries in parallel — timeout (default 5s) or any
        other DB error becomes 400; (4) row-limit check (default 1000 rows,
        checked BEFORE comparison to avoid a false-positive match) — 400 if
        exceeded; (5) `compareResults` — the main correctness check, several
        possible `failureReason` values; (6) if the results already matched,
        a SEPARATE structural pass can still flip a correct-looking answer to
        `failureReason: query_logic_mismatch` — first `validateRequiredOrderBy`
        (ORDER BY presence, always runs), then (only for `select`/`where`
        topics, or any task with `validationMode: strict`)
        `validateSqlStructure` (WHERE/DISTINCT/LIMIT/ORDER BY structural
        checks — deliberately biased toward false negatives over false
        positives, see `docs/check-answer-flow.md`).
      operationId: checkTaskAnswer
      x-visibility: authenticated
      x-writes-attempt: true
      x-updates-progress: true
      x-practice-area: sql
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userSql:
                  type: string
                  description: Must be a SELECT/WITH statement.
                sessionId:
                  type: integer
                  nullable: true
                  description: >-
                    Optional. See the endpoint description for the
                    non-obvious "unauthorized sessionId silently degrades to
                    no session" behavior.
              required: [userSql]
            example:
              userSql: "SELECT * FROM students;"
              sessionId: 42
      responses:
        "200":
          description: >-
            Always returned once SQL safety/schema/execution/row-limit checks
            pass — whether the answer was correct or not.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckAnswerResult" }
              examples:
                correct:
                  summary: "A) Correct answer (task 1, live-verified)"
                  value:
                    isCorrect: true
                    solutionHasJoin: false
                    userResult:
                      rows: [{ "id": 1, "first_name": "Ana", "last_name": "Kovač", "index_number": "2021001", "enrollment_year": 2021, "faculty_id": 1, "department_id": 1 }]
                      columns: ["id", "first_name", "last_name", "index_number", "enrollment_year", "faculty_id", "department_id"]
                      rowCount: 90
                    expectedRowCount: 90
                valueMismatch:
                  summary: "B) Incorrect — result mismatch, wrong values (task 15, live-verified: e.grade + 1)"
                  value:
                    isCorrect: false
                    failureReason: value_mismatch
                    userRowCount: 160
                    expectedRowCount: 160
                    sampleDifferences:
                      - row: 1
                        columns: ["grade"]
                        userValues: [11]
                        expectedValues: [10]
                      - row: 2
                        columns: ["grade"]
                        userValues: [11]
                        expectedValues: [10]
                      - row: 3
                        columns: ["grade"]
                        userValues: [11]
                        expectedValues: [10]
                    solutionHasJoin: true
                    userResult:
                      rows: [{ "first_name": "Ana", "last_name": "Kovač", "subject_name": "Databases", "grade": 11 }]
                      columns: ["first_name", "last_name", "subject_name", "grade"]
                      rowCount: 160
                rowCountMismatch:
                  summary: "Incorrect — row count mismatch (task 15, live-verified: added a restrictive WHERE)"
                  value:
                    isCorrect: false
                    failureReason: row_count_mismatch
                    userRowCount: 34
                    expectedRowCount: 160
                    solutionHasJoin: true
                    userResult:
                      rows: [{ "first_name": "Ana", "last_name": "Kovač", "subject_name": "Databases", "grade": 9 }]
                      columns: ["first_name", "last_name", "subject_name", "grade"]
                      rowCount: 34
                columnMismatch:
                  summary: "Incorrect — column count mismatch (task 15, live-verified: dropped the grade column)"
                  value:
                    isCorrect: false
                    failureReason: column_count_mismatch
                    userColumns: ["first_name", "last_name", "subject_name"]
                    expectedColumns: ["first_name", "last_name", "subject_name", "grade"]
                    missingColumns: ["grade"]
                    extraColumns: []
                    solutionHasJoin: true
                    userResult:
                      rows: [{ "first_name": "Ana", "last_name": "Kovač", "subject_name": "Databases" }]
                      columns: ["first_name", "last_name", "subject_name"]
                      rowCount: 160
                    expectedRowCount: 160
                orderMismatch:
                  summary: "Incorrect — right rows, wrong sequence (task 7, live-verified: solution has ORDER BY, user omitted it)"
                  value:
                    isCorrect: false
                    failureReason: order_mismatch
                    userRowCount: 48
                    expectedRowCount: 48
                    sampleDifferences:
                      - row: 1
                        columns: ["id", "name", "ects", "faculty_id", "department_id"]
                        userValues: [1, "Osnovi programiranja", 7, 1, 1]
                        expectedValues: [33, "Matematička analiza 1", 8, 5, 14]
                    solutionHasJoin: false
                    userResult:
                      rows: [{ "id": 1, "name": "Osnovi programiranja", "semester": 1, "ects": 7, "faculty_id": 1, "department_id": 1 }]
                      columns: ["id", "name", "semester", "ects", "faculty_id", "department_id"]
                      rowCount: 48
                queryLogicMismatch:
                  summary: >-
                    Incorrect — right rows, but a stricter structural check still rejects it
                    (task 1, live-verified: added "WHERE enrollment_year > 2000")
                  value:
                    isCorrect: false
                    failureReason: query_logic_mismatch
                    logicMismatchReason: extra_where_condition
                    logicMismatchHint: "Check your WHERE clause — you may have added a condition that makes the result too restrictive."
                    solutionHasJoin: false
                    userResult:
                      rows: [{ "id": 1, "first_name": "Ana", "last_name": "Kovač", "enrollment_year": 2021 }]
                      columns: ["id", "first_name", "last_name", "enrollment_year"]
                      rowCount: 90
                    expectedRowCount: 90
        "400":
          description: Missing userSql, dataset/session mismatch, a DB execution error/timeout, or the result exceeds the row limit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                missingSql:
                  summary: userSql missing (returned even when logged out — see description)
                  value: { error: "userSql is required." }
                datasetMismatch:
                  summary: Task's dataset does not match the resolved session's dataset
                  value: { error: 'This task belongs to the "academic" dataset but your session uses "cinema".' }
                timeout:
                  summary: Query exceeded the statement timeout
                  value: { error: "Your query exceeded the time limit of 5 seconds. Try a more specific query." }
                rowLimit:
                  summary: Result exceeds the row limit (checked before comparison)
                  value: { error: "Your query returns more than 1000 rows. The expected result has 90 rows. Add a more specific filter or a LIMIT clause." }
        "401":
          description: "Not authenticated. Only reachable when userSql was provided — see the endpoint description."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: "Not authenticated."
        "403":
          description: >-
            SQL failed the safety/schema-scope check, or the resolved
            session (one the caller actually owns) is completed or archived.
            NOT returned for a sessionId the caller doesn't own — see the
            endpoint description.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                unsafeSql:
                  summary: SQL failed the safety check
                  value: { error: "Only read-only SELECT queries are allowed. Detected: DELETE" }
                crossSchema:
                  summary: Cross-dataset schema reference
                  value: { error: 'Cross-dataset query not allowed. Your session only has access to the "academic" dataset.' }
                sessionCompleted:
                  summary: The (owned) resolved session is completed
                  value: { error: "This session is completed. Reopen it to continue." }
                sessionArchived:
                  summary: The (owned) resolved session is archived
                  value: { error: "This session is archived. Restore it from the sidebar to continue." }
        "404":
          description: No task with this id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: "Task not found"
...